# swagger.yaml
openapi: 3.0.0
paths:
  /api/auth/register:
    post:
      operationId: AuthController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: User registered successfully
        '409':
          description: Email already in use
      summary: Register a new user
      tags: &ref_0
        - auth
  /api/auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
      summary: Login with email and password
      tags: *ref_0
  /api/deals:
    get:
      operationId: DealsController_findAll
      parameters: []
      responses:
        '200':
          description: List of all deals
        '401':
          description: Unauthorized
      security:
        - JWT-auth: []
      summary: Get all open deals
      tags:
        - deals
  /api/applications:
    post:
      operationId: ApplicationsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationDto'
      responses:
        '201':
          description: ''
      security: &ref_1
        - JWT-auth: []
      summary: Submit an application for a deal
      tags: &ref_2
        - applications
  /api/applications/my:
    get:
      operationId: ApplicationsController_findMine
      parameters: []
      responses:
        '200':
          description: ''
      security: *ref_1
      summary: Get current user applications
      tags: *ref_2
info:
  title: Zenbit API
  description: Real estate investment platform API
  version: '1.0'
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    RegisterDto:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        email:
          type: string
          example: john@example.com
        password:
          type: string
          example: password123
          minLength: 6
      required:
        - name
        - email
        - password
    LoginDto:
      type: object
      properties:
        email:
          type: string
          example: john@example.com
        password:
          type: string
          example: password123
      required:
        - email
        - password
    CreateApplicationDto:
      type: object
      properties:
        dealId:
          type: string
          example: uuid-of-deal
        fullName:
          type: string
          example: John Doe
        phone:
          type: string
          example: '+1234567890'
        email:
          type: string
          example: john@example.com
      required:
        - dealId
        - fullName
        - phone
        - email